refactor: zst.py to raster_zonal_stats.py
